requirements.txt
Django==5.2.1
djangorestframework==3.16.1
django-filter==23.2
Pillow==10.4.0
reportlab==4.2.2
Faker==30.0.0
